+++ yash-env/src/job.rs
@@ -45,7 +45,7 @@ use crate::Env;
 use crate::io::Fd;
 use crate::semantics::{Divert, ExitStatus};
 use crate::signal;
-use crate::system::{Disposition, Sigaction, Sigmask, SigmaskOp, Signals, TcSetPgrp};
+use crate::system::{Disposition, Sigaction, Sigmask, SigmaskOp, Signals, Sigset as _, TcSetPgrp};
 use slab::Slab;
 use std::collections::HashMap;
 use std::iter::FusedIterator;
@@ -940,10 +940,13 @@ pub async fn tcsetpgrp_with_block<S>(system: &S, fd: Fd, pgid: Pid) -> crate::sy
 where
     S: Signals + Sigmask + TcSetPgrp + ?Sized,
 {
-    let mut old_mask = Vec::new();
+    let mut old_mask = S::Sigset::new();
 
     system
-        .sigmask(Some((SigmaskOp::Add, &[S::SIGTTOU])), Some(&mut old_mask))
+        .sigmask(
+            Some((SigmaskOp::Add, &S::Sigset::from_signals([S::SIGTTOU])?)),
+            Some(&mut old_mask),
+        )
         .await?;
 
     let result = system.tcsetpgrp(fd, pgid).await;
@@ -978,15 +981,14 @@ pub async fn tcsetpgrp_without_block<S>(system: &S, fd: Fd, pgid: Pid) -> crate:
 where
     S: Signals + Sigaction + Sigmask + TcSetPgrp + ?Sized,
 {
+    let sigttou = S::Sigset::from_signals([S::SIGTTOU])?;
+
     match system.sigaction(S::SIGTTOU, Disposition::Default) {
         Err(e) => Err(e),
         Ok(old_handling) => {
-            let mut old_mask = Vec::new();
+            let mut old_mask = S::Sigset::new();
             let result = match system
-                .sigmask(
-                    Some((SigmaskOp::Remove, &[S::SIGTTOU])),
-                    Some(&mut old_mask),
-                )
+                .sigmask(Some((SigmaskOp::Remove, &sigttou)), Some(&mut old_mask))
                 .await
             {
                 Err(e) => Err(e),